requirements.txt
pyrogram==2.0.106
discord.py==2.3.1
TgCrypto==1.2.5
pyyaml==6.0